engine/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
pymysql>=1.1.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
alembic>=1.13.1
aiomysql>=0.2.0
aiohttp>=3.9.1
pyarrow>=14.0.1
pandas>=2.1.4
boto3>=1.31.3